working-main/verify_api_skeleton.py
import requests
import json

BASE_URL = "http://127.0.0.1:5000"

def test_api():
    session = requests.Session()

    # 1. Login as Guardian (using demo credentials or creating new ones)
    # We'll use the demo data credentials from create_demo_content.py if possible, 
    # but I don't know them for sure. I'll rely on what I saw in seed_data or just Signup.
    
    # Actually, let's signup a temp guardian to be sure.
    print("--- 1. Signing up Temp Guardian ---")
    signup_data = {
        "name": "API Test Guardian",
        "email": "api_test_guardian@example.com",
        "password": "password123",
        "confirm_password": "password123",
        "phone": "1234567890"
    }
    # Note: The route is /signup/guardian/submit based on previous analysis or similar.
    # Wait, let me check app.py for the exact route.
    # I saw /create-user.html is the form. Action?
    # I'll check app.py content for routes.
    pass

if __name__ == "__main__":
    pass
