working-main/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
dnspython==2.6.1
Flask==3.1.2
Flask-Login==0.6.3
flask-cors==5.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pymongo==4.6.3
Werkzeug==3.1.3
google-generativeai>=0.3.0
